requirements.txt
streamlit
google-generativeai
python-dotenv
reportlab
plotly
pandas
PyPDF2
hashlib
datetime